Move shared socket code to module
