+++ src/sys/net/socket/tcp.rs
@@ -3,19 +3,14 @@ use crate::sys;
 use crate::api::fs::{FileIO, IO};
 use crate::sys::net::SocketStatus;
 
+use super::SOCKETS;
+use super::{random_port, wait};
+
 use alloc::vec;
 use bit_field::BitField;
-use lazy_static::lazy_static;
 use smoltcp::iface::SocketHandle;
-use smoltcp::iface::SocketSet;
 use smoltcp::socket::tcp;
-use smoltcp::time::Duration;
 use smoltcp::wire::IpAddress;
-use spin::Mutex;
-
-lazy_static! {
-    pub static ref SOCKETS: Mutex<SocketSet<'static>> = Mutex::new(SocketSet::new(vec![]));
-}
 
 fn tcp_socket_status(socket: &tcp::Socket) -> u8 {
     let mut status = 0;
@@ -27,14 +22,6 @@ fn tcp_socket_status(socket: &tcp::Socket) -> u8 {
     status.set_bit(SocketStatus::MayRecv as usize, socket.may_recv());
     status.set_bit(SocketStatus::CanRecv as usize, socket.can_recv());
     status
-}
-
-fn random_port() -> u16 {
-    49152 + sys::random::get_u16() % 16384
-}
-
-fn wait(duration: Duration) {
-    sys::time::sleep((duration.total_micros() as f64) / 1000000.0);
 }
 
 #[derive(Debug, Clone)]
@@ -123,15 +110,8 @@ impl TcpSocket {
     }
 
     pub fn accept(&mut self) -> Result<IpAddress, ()> {
-        //let timeout = 5.0;
-        //let started = sys::clock::realtime();
         if let Some((ref mut iface, ref mut device)) = *sys::net::NET.lock() {
             loop {
-                /*
-                if sys::clock::realtime() - started > timeout {
-                    return Err(());
-                }
-                */
                 let mut sockets = SOCKETS.lock();
                 iface.poll(sys::net::time(), device, &mut sockets);
                 let socket = sockets.get_mut::<tcp::Socket>(self.handle);
